Switch back to float

+++ inst/examples/sirs.cpp
@@ -1,6 +1,6 @@
 class sirs {
 public:
-  typedef double real_t;
+  typedef float real_t;
 
   struct init_t {
     real_t S0;
